Fix matching inaccuracy of TA code frames
The matching was ambiguous, causing the wrong filedata frames to be
mapped into a dataport for a specific TA. To replicate: assume we have
two TA's named "ta" and "ta2". Before this fix, finding all code pages
belonging to "ta" involved a search for `.name = "frame_ta`, but this
would match entries `.name = "frame_ta2` too. This is easily fixed by
including the `_group_bin` suffix in our search.